Prop Symbols Introduced
geojson.js cleared with relevant code moved to main.js.  Main.js works
correctly and displays asian census geojson dataset, but when assigning
attribute in main.js script it doesn't seem to be working as all 22
cities are displaying at once rather than just 20 and population values
for all years are displaying in console.log (pre-radius update).  Now,
with radius determination equations for prop. symbols, the symbols are
taking up and covering the entire map, so something must be wrong with
the function or the way that the dataset isn't being filtered into just
a single year of population data.